moved the db sub-command into a cmd/db module and updated the use statements to reflect the change
